Explicitly request use of conda in Jenkinsfile
This change causes the creation of a new conda environment used to host the build as per the `conda_packages` docs here: https://github.com/spacetelescope/jenkins_shared_ci_utils/blob/master/README.md and removes the `with_env` commands which are then no longer necessary.

The error noted in issue #403 was caused by the Jenkinsfile not requesting the use of conda (by populating the `conda_packages` list for a given build configuration). A leftover pre-installed conda in the job's container was used by default, which allowed the `conda` commands specified in the build configuration to execute, though on a conda base environment, rather than a new target environment, which is not desired.

As an additional safeguard, future build containers on the Jenkins host will do away with the built-in conda installation to make the behavior more consistent.